add persistent resize mode
